fix (i18n): rename i18n root folder
