+++ .gitignore
@@ -21,7 +21,7 @@ web-build/
 
 # Firebase Service Account
 credentials/
-*.json
+
 
 # debug
 npm-debug.*
