+++ components/services/localization/LanguageModal.tsx
@@ -1,3 +1,9 @@
+/////////////////////////////////// LanguageModal.tsx //////////////////////////////////
+
+// This component is used to show a modal to select the language
+
+/////////////////////////////////////////////////////////////////////////////////////////
+
 import {
   View,
   Text,
