+++ components/services/localization/languages.ts
@@ -1,3 +1,9 @@
+/////////////////////////// languages.ts /////////////////////////////
+
+// This file contains the languages supported by the app
+
+////////////////////////////////////////////////////////////////////////
+
 export const LANGUAGE_OPTIONS = [
   {
     code: "en",
