+++ validation/noteSchemas.ts
@@ -5,7 +5,7 @@
 //////////////////////////////////////////////////////////////
 
 import { z } from "zod";
-import i18n from "../components/services/lacalization/i18n";
+import i18n from "../components/services/localization/i18n";
 
 //////////////////////////////////////////////////////////////
