+++ validation/timeTrackingStateSchemas.ts
@@ -6,7 +6,7 @@
 
 import { z } from "zod";
 import { ProjectState } from "../components/TimeTrackingState";
-import i18n from "../components/services/lacalization/i18n";
+import i18n from "../components/services/localization/i18n";
 
 ////////////////////////////////////////////////////////////////////
